Mise à jour de la suppression des sauces

+++ back/controllers/sauces.js
@@ -1,11 +1,15 @@
 const Sauce = require("../models/sauces");
 const fs = require("fs");
 
+// Affichage des sauces
+
 exports.getSauces = (req, res, next) => {
   Sauce.find()
     .then((sauces) => res.status(200).json(sauces))
     .catch((error) => res.status(400).json({ error }));
 };
+
+// Création des sauces
 
 exports.createSauce = (req, res, next) => {
   const sauceObject = JSON.parse(req.body.sauce);
@@ -29,11 +33,15 @@ exports.createSauce = (req, res, next) => {
     });
 };
 
+// Affichage d'une sauce
+
 exports.getOneSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauces) => res.status(200).json(sauces))
     .catch((error) => res.status(400).json({ error }));
 };
+
+// Modification des sauces
 
 exports.modifySauce = (req, res, next) => {
   const sauceObject = req.file
@@ -64,8 +72,25 @@ exports.modifySauce = (req, res, next) => {
     });
 };
 
+// Suppression des sauces
+
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Sauce deleted!" }))
-    .catch((error) => res.status(400).json({ error }));
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Sauce deleted!" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
